fix(refactoring): moving brandAsset constant to payment methods file
